+++ .travis.yml
@@ -1,6 +1,6 @@
 language: node_js
 node_js:
-  - "11"
+  - "13"
 cache: yarn
 after_success:
   - npm run coverage
