+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.1.1] - 2019-12-29
+
+### Changed
+* moved API docs to readme
+* bumped devDependencies
+* moved `src/index.js` to be just at `index.js`
+* changed license from ISC to BSD-3
+
 ## [0.1.0] - 2018-12-29
 
 ### Added
