+++ package.json
@@ -7,7 +7,7 @@
     "AUTHORS",
     "LICENSE",
     "README.md",
-    "src/"
+    "index.js"
   ],
   "repository": {
     "type": "git",
@@ -16,10 +16,7 @@
   "homepage": "https://github.com/rpearce/simple-uniqueid",
   "bugs": "https://github.com/rpearce/simple-uniqueid/issues",
   "author": "Robert Pearce <me@robertwpearce.com>",
-  "contributors": [
-    "Robert Pearce <me@robertwpearce.com> (https://robertwpearce.com)"
-  ],
-  "license": "ISC",
+  "license": "BSD-3",
   "keywords": [
     "uniqueid",
     "id",
@@ -34,9 +31,9 @@
     "contributors:add": "all-contributors add",
     "contributors:generate": "all-contributors generate",
     "coverage": "jest --coverage --coverageReporters=text-lcov | coveralls",
-    "lint": "eslint --ext .js .",
-    "prepublishOnly": "npm run lint",
-    "test": "npm run lint && jest --coverage --no-cache"
+    "lint": "eslint .",
+    "prepublishOnly": "run-p lint test",
+    "test": "jest"
   },
   "jest": {
     "clearMocks": true,
@@ -45,10 +42,11 @@
     ]
   },
   "devDependencies": {
-    "all-contributors-cli": "^5.4.1",
-    "coveralls": "^3.0.2",
-    "eslint": "^5.11.1",
-    "eslint-plugin-jest": "^22.1.2",
-    "jest": "^23.6.0"
+    "all-contributors-cli": "^6.11.2",
+    "coveralls": "^3.0.9",
+    "eslint": "^6.8.0",
+    "eslint-plugin-jest": "^23.2.0",
+    "jest": "^24.9.0",
+    "npm-run-all": "^4.1.5"
   }
 }
